Fronted/Seguros_Confianza/Backend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
filelock==3.16.1
mysqlclient==2.2.4
platformdirs==4.3.6
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.27.0
